Fix remove test
