+++ src/modules/AddTest.js
@@ -16,3 +16,22 @@ export const removeTask = (tasks, index) => {
     i += 1;
   }
 };
+
+export const viewTask = (tasks) => {
+  const container = document.querySelector('.to-do-list');
+  while (container.children[2]) {
+    container.removeChild(container.children[2]);
+  }
+  tasks.sort((a, b) => a.index - b.index);
+  tasks.forEach((task) => {
+    const listItem = document.createElement('li');
+    listItem.className = 'list-item';
+    listItem.innerHTML = `
+      <input class="check-box" type="checkbox" ${task.completed ? 'checked' : 'unchecked'}>
+      <p class="description">${task.description}</p>
+      `;
+    container.appendChild(listItem);
+  });
+
+  localStorage.setItem('toDoList', JSON.stringify(tasks));
+};
