+++ src/modules/add.test.js
@@ -1,29 +1,21 @@
 /** @jest-environment jsdom */
 
-import {addTask , removeTask} from "./AddTest";
-
-// const addTask = require('./AddTest.js');
-// const removeTask = require('./AddTest.js');
+import {addTask , removeTask, viewTask} from "./AddTest";
 
 describe('Adding or removing items test', () => {
-  document.body.innerHTML += '<ul class="to-do-list"> </ul>';
-  const container = document.querySelector('.to-do-list');
-  const tasks = [];
-  const listItem = document.createElement('li');
-  listItem.className = 'list-item';
+  let tasks;
+
+  beforeEach (() => {
+    tasks = [];
+    document.body.innerHTML += '<ul class="to-do-list"> </ul>';
+  })
 
   test('Adding one item to the list', () => {
     tasks.push(addTask('Testing', tasks.length));
 
     expect(tasks).toEqual([{ description: 'Testing', completed: false, index: 0 }]);
-    
-    tasks.forEach((task) => {
-      listItem.innerHTML = `
-        <input class="check-box" type="checkbox" ${task.completed ? 'checked' : 'unchecked'}>
-        <p class="description">${task.description}</p>
-        `;
-      container.appendChild(listItem);
-    });
+
+    viewTask(tasks);
 
     const list = document.querySelectorAll('.list-item');
 
@@ -31,20 +23,12 @@ describe('Adding or removing items test', () => {
   });
 
   test('Removing one item from the list', () => {
+    tasks.push(addTask('Testing', tasks.length));
+    tasks.push(addTask('Testing', tasks.length));
     removeTask(tasks, 0);
 
-    tasks.forEach((task) => {
-      listItem.innerHTML = `
-        <input class="check-box" type="checkbox" ${task.completed ? 'checked' : 'unchecked'}>
-        <p class="description">${task.description}</p>
-        `;
-      container.appendChild(listItem);
-    });
-    
-    const list = document.querySelectorAll('.list-item');
-    
-    console.log(list)
+    viewTask(tasks);
 
-    expect(list).toHaveLength(0);
+    expect(tasks).toHaveLength(1);
   });
 });
